Comment out file owner and group specification
In the task for transferring Iris installation files, the lines specifying the owner and group have been commented out. The intent of these changes is to use default values for these attributes, rather than explicitly setting them. The functionality for the ability to specify these is preserved, in case future requirements necessitate their usage.